JAVA FX RFEMOVED  UNSTABLE

+++ src/main/java/com/allinweb/ch/component/scene/ARExcelFileScene.java
@@ -51,12 +51,12 @@ public class ARExcelFileScene extends ARScene {
     }
 
     @Override
-    public Double getSceneHeight() {
+    public int getSceneHeight() {
         return SCENE_HEIGHT;
     }
 
     @Override
-    public Double getSceneWidth() {
+    public int getSceneWidth() {
         return SCENE_WIDTH;
     }
