+++ src/main/java/com/allinweb/ch/component/scene/ARInfoScene.java
@@ -48,12 +48,12 @@ public class ARInfoScene extends ARScene {
     }
 
     @Override
-    public Double getSceneHeight() {
+    public int getSceneHeight() {
         return SCENE_HEIGHT;
     }
 
     @Override
-    public Double getSceneWidth() {
+    public int getSceneWidth() {
         return SCENE_WIDTH;
     }
