+++ src/main/java/com/allinweb/ch/component/scene/ARViewBotJobListScene.java
@@ -60,12 +60,12 @@ public class ARViewBotJobListScene extends ARScene {
     }
 
     @Override
-    public Double getSceneHeight() {
+    public int getSceneHeight() {
         return SCENE_HEIGHT;
     }
 
     @Override
-    public Double getSceneWidth() {
+    public int getSceneWidth() {
         return SCENE_WIDTH;
     }
